Draw all components after map image (#15)
This change fixes a bug where only the map title was drawn once the map
image was loaded. This meant that the pins would be drawn first and would
be drawn over by the map image afterwards. Instead, now everything is draw
once we know that the main map image has been completely drawn as it will
be overlayed with everything else.