Adjust font based on image dimensions (#21)
This change adjusts the text's position and dimensions based on the
dimensions of the paired image. For the x co-ordinate, the width of the
text is taken into consideration to center the text perfectly in relation
to the image. For the font height, different logic is applied based on
component but is essentially a percentage based on the image height.