Add circle for interaction
This change adds the initial circle below the map pin intended to be for
interaction.  This has been done by redesigning the dimensions to include
different types of values for different shapes as well as adding a circle
svg component. The intention is to have a colour passed in and work with
that but for now, we are using a default colour. A TODO has been added to
add the future feature.

Close #25