Add coordinate space for map (#41)
This change allows users of the library to define a coordinate space
(width and height) on which their pins are based. This means that when
they can define a dynamic svg component and the map will automatically
scale the provided components. The demo was also updated to allow playing
around with the svg component values