Allow users to pass pin dimensions
This change allows users to pass in custom dimensions for each pin to draw
on the map as it could have different height and width rather than being a
square. In addition, scaling for this with map dimensions has also been
implemented. Lastly, the demo was changed to allow incorporation of the
width and height of each pin to allow users to experiment

Close #40