Add demo feature
This change allows user to run a script to run the demo on a local
http server. To do this, the webpack.config.js to generate the min.js
file for the library in the demo folder as well as the dist folder. This is
so that we can serve the demo folder and have it available in there.
Scripts have also been created to build and serve the demo